update scoring server \'s README
