+++ services/rewarding/app.py
@@ -1,448 +1,872 @@
-from fastapi import FastAPI
-import uvicorn
-from redteam_core import (
-    challenge_pool,
-    constants,
-)
-from redteam_core.validator.miner_manager import (
-    ChallengeRecord,
-    MinerManager,
-    ScoringLog
-)
-from redteam_core.validator.storage_manager import StorageManager
-from cryptography.fernet import Fernet
-import requests
 import argparse
+import datetime
+import json
+import os
 import threading
 import time
-import copy
-import datetime
+import traceback
+from typing import Annotated, Union
+
 import bittensor as bt
-import os
+import requests
+import uvicorn
+from fastapi import Body, FastAPI
 from prometheus_fastapi_instrumentator import Instrumentator
 
-class ChallengeRecord(BaseModel):
-    point: float = 0
-    score: float = 0
-    date: str = datetime.datetime.now(datetime.timezone.utc).strftime("%Y-%m-%d")
-    scored_date: Optional[str] = None
-    docker_hub_id: Optional[str] = None
-    uid: Optional[int] = None
-    ss58_address: Optional[str] = None
+from neurons.validator.validator import Validator
+from redteam_core import constants
+from redteam_core.common import get_config
+from redteam_core.validator.models import (
+    ComparisonLog,
+    MinerChallengeCommit,
+    ScoringLog,
+)
+
+REWARD_APP_HOTKEY = os.getenv("REWARD_APP_HOTKEY")
+REWARD_APP_UID = -1
 
 
-class ScoringLog(BaseModel):
-    score: float
-    miner_input: Optional[dict] = None
-    miner_output: Optional[dict] = None
-    error: Optional[str] = None
-    baseline_score: Optional[float] = None
-
-def get_args():
+def get_reward_app_config() -> bt.Config:
     parser = argparse.ArgumentParser()
-    parser.add_argument("--port", type=int, default=10001)
-    parser.add_argument("--netuid", type=int, default=61)
-    parser.add_argument("--network", type=str, default="finney")
-    parser.add_argument("--reward_app_key", type=str, default=os.getenv("REWARD_APP_KEY"))
-    parser.add_argument("--cache_dir", type=str, default="cache_reward_app")
-    parser.add_argument("--hf_repo_id", type=str, default=os.getenv("HF_REPO_ID"))
-    parser.add_argument("--reward_app_ss58", type=str, default=os.getenv("REWARD_APP_SS58"))
-    args = parser.parse_args()
-    return args
+    parser.add_argument("--reward_app.port", type=int, default=47920)
+    parser.add_argument("--reward_app.epoch_length", type=int, default=60)
+    config = get_config(parser)
+    return config
 
-class RewardApp:
-    def __init__(self, args):
-        self.args = args
-        self.REWARD_APP_KEY = args.reward_app_key
-        self.REWARD_APP_SERVER_SS58_ADDRESS = args.reward_app_ss58
-        self.REWARD_APP_SERVER_UID = -1
-        # self.active_challenges = challenge_pool.ACTIVE_CHALLENGES
-        self.subtensor = bt.subtensor(network=args.network)
-        self.metagraph = self.subtensor.metagraph(args.netuid, lite=True)
-        self.miner_managers = {}
-        self.smooth_transition_challenge()
-        self._init_challenge_records_from_subnet()
-        self.submission_scoring_logs = self.fetch_submission_scoring_logs(list(self.active_challenges.keys()))
-        self.previous_submission_scoring_logs = copy.deepcopy(self.submission_scoring_logs)
 
-        self.is_scoring_done = {
+class RewardApp(Validator):
+    """
+    A special validator that focuses on centralized scoring for the network.
+    This validator:
+    1. Does not participate in querying miners or setting weights
+    2. Maintains state like regular validators
+    3. Scores miner commits retrieved from storage
+    4. Provides API endpoints for direct access to scoring results
+    """
+
+    def __init__(self, config: bt.Config):
+        """
+        Initialize the reward app with validator capabilities.
+
+        Args:
+            config (bt.Config): Bittensor configuration object
+
+        State Management:
+            - validators_miner_commits: Stores current miner commits from all validators
+            - miner_commits: Aggregated miner commits from all validators
+            - miner_commits_cache: Quick lookup cache mapping challenge_name---encrypted_commit to commit
+            - scoring_results: Cache for scored docker_hub_ids with their scoring and comparison logs
+            - is_scoring_done: Tracks scoring completion status for each challenge
+        """
+        super().__init__(config)
+        # Initialize validator state, stores current miner commits from all validators
+        self.validators_miner_commits: dict[
+            tuple[int, str], dict[tuple[int, str], dict[str, MinerChallengeCommit]]
+        ] = {}
+        # Daily miner commits will now be aggregated in self.miner_commits
+        self.miner_commits: dict[tuple[int, str], dict[str, MinerChallengeCommit]] = {}
+        # Quick lookup for miner commits by encrypted_commit, this has no new information, just a cache
+        self.miner_commits_cache: dict[str, MinerChallengeCommit] = {}
+        # Cache for scored docker_hub_ids, map from challenge_name to docker_hub_id to the coresponding "scoring_logs" and "comparison_logs"
+        self.scoring_results: dict[
+            str, dict[str, Union[list[ScoringLog], dict[str, ComparisonLog]]]
+        ] = self._fetch_centralized_scoring(list(self.active_challenges.keys()))
+        # Sync the cache from scoring results retrieved from storage upon initialization
+        self._sync_scoring_results_from_storage_to_cache()
+
+        # Initialize scoring completion flags
+        # This is used to check if the scoring for a challenge is done
+        # Set to False when new miner commits are retrieved and True when all miner commits are scored and compared at scoring hour
+        self.is_scoring_done: dict[str, bool] = {
             challenge_name: False for challenge_name in self.active_challenges.keys()
         }
-        self.miner_submit = {}
-        self.mapping_docker_id_miner_id = {}
-        self.scoring_dates: list[str] = []
 
-        self.storage_manager = StorageManager(
-            cache_dir=self.args.cache_dir,
-            hf_repo_id=self.args.hf_repo_id,
-            sync_on_init=False
+        # Initialize FastAPI app (May change this to use bt.axon in the future)
+        self.app = FastAPI()
+        self.app.add_api_route(
+            "/get_scoring_result",
+            self.get_scoring_result,
+            methods=["POST"],
+        )
+        Instrumentator().instrument(self.app).expose(self.app)
+        # Run FastAPI server in a separate thread
+        self.server_thread = threading.Thread(
+            target=uvicorn.run,
+            kwargs={
+                "app": self.app,
+                "host": "0.0.0.0",
+                "port": self.config.reward_app.port,
+                "log_level": "debug",
+            },
+            daemon=True,  # Ensures the thread stops when the main process exits
+        )
+        self.server_thread.start()
+        bt.logging.info(f"FastAPI server is running on port {self.config.reward_app.port}!")
+
+    def setup_bittensor_objects(self):
+        bt.logging.info("Setting up Bittensor objects.")
+        self.wallet = bt.wallet(config=self.config)
+        bt.logging.info(f"Wallet: {self.wallet}")
+        self.subtensor = bt.subtensor(config=self.config)
+        bt.logging.info(f"Subtensor: {self.subtensor}")
+        self.dendrite = bt.dendrite(wallet=self.wallet)
+        bt.logging.info(f"Dendrite: {self.dendrite}")
+        self.metagraph = self.subtensor.metagraph(self.config.netuid)
+        bt.logging.info(f"Metagraph: {self.metagraph}")
+
+        if REWARD_APP_HOTKEY != self.wallet.hotkey.ss58_address:
+            bt.logging.error(
+                f"Reward app hotkey {REWARD_APP_HOTKEY} does not match wallet hotkey {self.wallet.hotkey.ss58_address}"
+            )
+            exit()
+        else:
+            self.hotkey = REWARD_APP_HOTKEY
+            self.uid = REWARD_APP_UID
+            bt.logging.success(
+                f"Reward app initialized with hotkey: {self.hotkey}, uid: {self.uid}"
+            )
+
+    # MARK: Validation Loop
+    def forward(self):
+        # 1. Update subnet commits state
+        # Update commits from all validators
+        self._update_validators_miner_commits()
+        # Update (aggregate) miner commits
+        self._update_miner_commits()
+        # Get revealed commits
+        revealed_commits = self.get_revealed_commits()
+
+        # Update miner infos
+        for challenge, challenge_manager in self.challenge_managers.items():
+            if challenge not in revealed_commits:
+                continue
+            challenge_manager.update_miner_infos(
+                miner_commits=revealed_commits.get(challenge, [])
+            )
+
+            # Flag the challenge as not done scoring if there are still unscored submissions
+            if revealed_commits[challenge]:
+                self.is_scoring_done[challenge] = False
+
+        # 2. Score and compare miner commits for each challenge
+        for challenge in revealed_commits:
+            if revealed_commits[challenge]:
+                # Score and compare new commits
+                self._score_and_compare_new_miner_commits(
+                    challenge=challenge,
+                    revealed_commits_list=revealed_commits[challenge],
+                )
+
+                # Update cache
+                for commit in revealed_commits[challenge]:
+                    self.scoring_results.setdefault(challenge, {})[
+                        commit.docker_hub_id
+                    ] = {
+                        "scoring_logs": commit.scoring_logs,
+                        "comparison_logs": commit.comparison_logs,
+                    }
+
+                bt.logging.info(
+                    f"[CENTRALIZED SCORING] Scoring for challenge: {challenge} has been completed"
+                )
+
+                # Store commits and scoring cache from this challenge
+                self._store_miner_commits(
+                    miner_commits={challenge: revealed_commits[challenge]}
+                )
+                self._store_centralized_scoring(challenge_name=challenge)
+
+        # Store reward app state, this can be viewed by other validators, so we need to make it public view
+        self.storage_manager.update_validator_state(
+            data=self.export_state(public_view=True), async_update=True
         )
 
-        self.sync_metagraph_thread = threading.Thread(target=self._sync_metagraph, daemon=True).start()
-        self.scoring_thread = threading.Thread(target=self.reward_submission, daemon=True).start()
-        self.finalize_challenge_result_thread = threading.Thread(target=self.get_challenge_result_of_today, daemon=True).start()
+        # 3. Finalize validator's daily state
+        # Get current time info
+        today = datetime.datetime.now(datetime.timezone.utc)
+        today_key = today.strftime("%Y-%m-%d")
+        current_hour = today.hour
 
-        self.app = FastAPI()
-        self.app.add_api_route("/get_scoring_logs", self.get_scoring_logs, methods=["GET"])
-        Instrumentator().instrument(self.app).expose(self.app)
+        validate_scoring_hour = current_hour >= constants.SCORING_HOUR
+        validate_scoring_date = today_key not in self.scoring_dates
 
-    def smooth_transition_challenge(self):
-        # TODO: Remove this next update
-        """
-        Smooth transition challenge from old to new challenge
-        """
-        all_challenges = copy.deepcopy(challenge_pool.ACTIVE_CHALLENGES)
+        if validate_scoring_hour and validate_scoring_date:
+            # At this point, all commits should be scored and compared against previous unique commits already, we now need to compare new commits with each other
+            for challenge in revealed_commits:
+                if revealed_commits[challenge]:
+                    self._compare_miner_commits(
+                        challenge=challenge,
+                        revealed_commits_list=revealed_commits[challenge],
+                        compare_with_each_other=True,
+                    )
 
-        if datetime.datetime.now(datetime.timezone.utc) <= datetime.datetime(2025, 1, 15, 14, 0, 0, 0, datetime.timezone.utc):
-            all_challenges.pop("response_quality_adversarial_v2", None)
-            all_challenges.pop("response_quality_ranker_v2", None)
-            all_challenges.pop("webui_auto", None)
+            # Update scores and penalties to challenge manager and mark challenge as done
+            for challenge in revealed_commits:
+                self.challenge_managers[challenge].update_miner_scores(
+                    miner_commits=revealed_commits[challenge]
+                )
+                self.is_scoring_done[challenge] = True
+
+                # Store commits and scoring cache from this challenge
+                self._store_miner_commits(
+                    miner_commits={challenge: revealed_commits[challenge]}
+                )
+                self._store_centralized_scoring(challenge_name=challenge)
+
+            self.scoring_dates.append(today_key)
+
+            # Store reward app state, this can be viewed by other validators, so we need to make it public view
+            self.storage_manager.update_validator_state(
+                data=self.export_state(public_view=True), async_update=True
+            )
         else:
-            all_challenges.pop("response_quality_adversarial", None)
-            all_challenges.pop("response_quality_ranker", None)
+            bt.logging.info(
+                f"[CENTRALIZED FORWARD] Not time to finalize daily result. Hour: {current_hour}, Date: {today_key}"
+            )
 
-        self.active_challenges = all_challenges
-
-        for challenge in self.active_challenges.keys():
-            if challenge not in self.miner_managers:
-                self.miner_managers[challenge] = MinerManager(
-                    challenge_name=challenge,
-                    challenge_incentive_weight=self.active_challenges[challenge]["challenge_incentive_weight"],
-                    metagraph=self.metagraph
-                )
-
-    def reward_submission(self):
-        """Background thread to reward submission.
-        1. Fetch miner submit
-        2. Group miner submit by challenge
-        3. Run challenges
-        4. Save submission scoring logs
-        5. If no new submission, sleep for 60 seconds
+    def _score_and_compare_new_miner_commits(
+        self, challenge: str, revealed_commits_list: list[MinerChallengeCommit]
+    ):
         """
-        while True:
-            self.smooth_transition_challenge()
-            print("Active challenges: ", self.active_challenges.keys())
-            self._fetch_miner_submit()
-            grouped_miner_submit, self.mapping_docker_id_miner_id = self.group_miner_submit_by_challenge(self.miner_submit)
-            new_submission_scoring_logs = self.run_challenges(grouped_miner_submit)
-            is_updated = self.save_submission_scoring_logs(new_submission_scoring_logs)
-            if not is_updated:
-                print("[INFO] No new submission, sleeping for 60 seconds")
-                time.sleep(60)
+        Score and do comparison for new miner commits for a specific challenge.
+        The default comparing behaviour is to compare new commits with previous unique commits only, not with each other since we don't have new commits for the whole day yet
 
-    def get_challenge_result_of_today(self):
-        while True:
-            today = datetime.datetime.now(datetime.timezone.utc)
-            today_key = today.strftime("%Y-%m-%d")
-            current_hour = today.hour
-            validate_scoring_hour = current_hour >= constants.SCORING_HOUR
-            validate_scoring_date = today_key not in self.scoring_dates
-            # Validate if scoring is due
-            if validate_scoring_hour and validate_scoring_date:
-                bt.logging.info(f"[SCORING] Scoring for {today_key} is due")
-                for challenge_name in self.miner_managers:
-                    if self.is_scoring_done.get(challenge_name, False):
-                        scoring_logs = []
-                        submission_scoring_logs = self.submission_scoring_logs[challenge_name]
-                        for docker_hub_id, logs in submission_scoring_logs.items():
-                            for log in logs:
-                                if docker_hub_id in self.mapping_docker_id_miner_id:
-                                    log = self._normalize_log(log)
-                                    scoring_logs.append(
-                                        ScoringLog(
-                                            uid=self.mapping_docker_id_miner_id[docker_hub_id]["uid"], # TODO: Change to log["uid"] in next update
-                                            ss58_address=self.mapping_docker_id_miner_id[docker_hub_id]["ss58_address"], # TODO: Change to log["ss58_address"] in next update
-                                            score=log["score"],
-                                            miner_input=log.get("miner_input"),
-                                            miner_output=log.get("miner_output"),
-                                            miner_docker_image=docker_hub_id,
-                                            error=log.get("error"),
-                                            baseline_score=log.get("baseline_score")
-                                        )
-                                )
-                        self.miner_managers[challenge_name].update_scores(scoring_logs)
-                        self._store_challenge_records()
-                        bt.logging.info(f"[SCORING] Scoring for challenge: {challenge_name} has been completed for {today_key}")
+        Args:
+            challenge (str): Challenge name
+            revealed_commits_list (list[MinerChallengeCommit]): List of new commits to score
 
-                        if all(self.is_scoring_done.get(challenge_name, False) for challenge_name in self.active_challenges.keys()):
-                            bt.logging.info(f"[SCORING] All tasks: Scoring completed for {today_key}")
-                            self.scoring_dates.append(today_key)
+        Process:
+        1. Look up cached results for already scored commits, use cached results for already scored commits
+        2. Gather unique commits from challenge manager for comparison
+        3. Retrieve cached data for reference commits
+        4. Run challenge controller with:
+           - New commits to be scored
+           - Reference commits for comparison
 
-            time.sleep(100)
+        """
+        if challenge not in self.active_challenges:
+            return
 
-    def run_challenges(self, docker_images_by_challenge: dict):
-        new_submission_scoring_logs = {}
-        for challenge_name, challenge_info in self.active_challenges.items():
-            try:
-                if challenge_name not in self.submission_scoring_logs:
-                    self.submission_scoring_logs[challenge_name] = {}
-                if challenge_name not in new_submission_scoring_logs:
-                    new_submission_scoring_logs[challenge_name] = {}
-                not_scored_submissions = [docker_hub_id for docker_hub_id in docker_images_by_challenge.get(challenge_name, []) if docker_hub_id not in self.submission_scoring_logs.get(challenge_name)]
-                not_scored_submissions = list(set(not_scored_submissions))
-                not_scored_uid_ss58_address_pairs = [(self.mapping_docker_id_miner_id[docker_hub_id]["uid"], self.mapping_docker_id_miner_id[docker_hub_id]["ss58_address"]) for docker_hub_id in not_scored_submissions]
-                if len(not_scored_submissions) == 0:
-                    self.is_scoring_done[challenge_name] = True
+        bt.logging.info(
+            f"[CENTRALIZED SCORING] Scoring miner commits for challenge: {challenge}"
+        )
+
+        if not revealed_commits_list:
+            bt.logging.info(
+                f"[CENTRALIZED SCORING] No commits for challenge: {challenge}, skipping"
+            )
+            return
+
+        # 1. Look up cached results for already scored commits, use cached results for already scored commits
+        # Also construct input seeds for new commits, this will be using input from commits that in the same revealed list for comparison
+        # We do this since commits being in the same revealed list means that they will be scored in same day
+        new_commits: list[MinerChallengeCommit] = []
+        seed_inputs: list[dict] = []
+
+        input_seed_hashes_set: set[str] = set()
+        for commit in revealed_commits_list:
+            if commit.docker_hub_id in self.scoring_results.setdefault(challenge, {}):
+                # Use results for already scored commits
+                cached_result = self.scoring_results[challenge][commit.docker_hub_id]
+                commit.scoring_logs = cached_result["scoring_logs"]
+                commit.comparison_logs = cached_result["comparison_logs"]
+
+                # Add input seed hash to set
+                for scoring_log in commit.scoring_logs:
+                    if (
+                        scoring_log.input_hash
+                        and scoring_log.input_hash not in input_seed_hashes_set
+                    ):
+                        input_seed_hashes_set.add(scoring_log.input_hash)
+                        seed_inputs.append(scoring_log.miner_input)
+            else:
+                new_commits.append(commit)
+
+        if not new_commits:
+            # No new commits to score, skip
+            bt.logging.info(
+                f"[CENTRALIZED SCORING] No new commits to score for challenge: {challenge}, skipping"
+            )
+            return
+        else:
+            bt.logging.info(
+                f"[CENTRALIZED SCORING] {len(new_commits)} new commits to score for challenge: {challenge}"
+            )
+
+        bt.logging.info(
+            f"[CENTRALIZED SCORING] Running controller for challenge: {challenge}"
+        )
+
+        # 2. Gather comparison inputs
+        # Get unique commits for the challenge (the "encrypted_commit"s)
+        unique_commits = self.challenge_managers[challenge].get_unique_commits()
+        # Get unique solutions 's cache key
+        unique_commits_cache_keys = [
+            self.storage_manager.hash_cache_key(unique_commit)
+            for unique_commit in unique_commits
+        ]
+        # Get commit 's cached data from storage
+        unique_commits_cached_data: list[MinerChallengeCommit] = []
+        challenge_local_cache = self.storage_manager._get_cache(challenge)
+        if challenge_local_cache:
+            unique_commits_cached_data_raw = [
+                challenge_local_cache.get(unique_commit_cache_key)
+                for unique_commit_cache_key in unique_commits_cache_keys
+            ]
+
+            unique_commits_cached_data = []
+            for commit in unique_commits_cached_data_raw:
+                if not commit:
                     continue
-                else:
-                    self.is_scoring_done[challenge_name] = False
-                controller = challenge_info["controller"](
-                    challenge_name=challenge_name,
-                    miner_docker_images=not_scored_submissions,
-                    uid_ss58_address_pairs=not_scored_uid_ss58_address_pairs,
-                    challenge_info=challenge_info
-                )
-                print(f"[CHALLENGE] Challenge {challenge_name} has been started")
-                logs = controller.start_challenge()
-                for log in logs:
-                    log = self._normalize_log(log)
-                    miner_docker_image = log["miner_docker_image"]
-                    if miner_docker_image not in self.submission_scoring_logs[challenge_name]:
-                        self.submission_scoring_logs[challenge_name][miner_docker_image] = []
-                    self.submission_scoring_logs[challenge_name][miner_docker_image].append(log)
-
-                    if miner_docker_image not in new_submission_scoring_logs[challenge_name]:
-                        new_submission_scoring_logs[challenge_name][miner_docker_image] = []
-                    new_submission_scoring_logs[challenge_name][miner_docker_image].append(log)
-            except Exception as e:
-                print(f"[ERROR] Error running challenge {challenge_name}: {e}")
-                # self.is_scoring_done[challenge_name] = True
-        return new_submission_scoring_logs
-
-    def group_miner_submit_by_challenge(self, miner_submit: dict):
-        docker_images_by_challenge = {}
-        mapping_docker_id_miner_id = {}
-        for miner_address, challenges in miner_submit.items():
-            for challenge_name, commit_data in challenges.items():
-                if challenge_name not in self.active_challenges:
-                    continue
-                if challenge_name not in docker_images_by_challenge:
-                    docker_images_by_challenge[challenge_name] = []
                 try:
-                    if "docker_hub_id" in commit_data:
-                        docker_hub_id = commit_data["docker_hub_id"]
-                    elif not commit_data.get("commit") and commit_data.get("key") and time.time() - commit_data["commit_timestamp"] > 24 * 60 * 60 and constants.is_commit_on_time(commit_data["commit_timestamp"]):
-                        f = Fernet(commit_data["key"])
-                        commit = f.decrypt(commit_data["encrypted_commit"]).decode()
-                        docker_hub_id = commit.split("---")[1]
-                    else:
-                        docker_hub_id = commit_data["commit"].split("---")[1]
+                    validated_commit = MinerChallengeCommit.model_validate(commit)
+                    unique_commits_cached_data.append(validated_commit)
+                except Exception:
+                    bt.logging.warning(
+                        f"[CENTRALIZED SCORING] Failed to validate cached commit {commit} for challenge {challenge}: {traceback.format_exc()}"
+                    )
+                    continue
 
-                    current_submit = mapping_docker_id_miner_id.get(docker_hub_id, {})
-                    # If current mapping is not exist or current mapping is exist but new commit timestamp is older than current commit timestamp, update mapping"""
-                    if not current_submit or (current_submit and current_submit["commit_timestamp"] > commit_data["commit_timestamp"]):
-                        mapping_docker_id_miner_id[docker_hub_id] = {
-                            "uid": commit_data["uid"],
-                            "ss58_address": commit_data["ss58_address"],
-                            "commit_timestamp": commit_data["commit_timestamp"]
-                        }
-                    docker_images_by_challenge[challenge_name].append(docker_hub_id)
-                except Exception as e:
-                    print(f"[ERROR] Error getting docker hub id: {e}")
-        return docker_images_by_challenge, mapping_docker_id_miner_id
+        # 3. Run challenge controller
+        bt.logging.info(
+            f"[CENTRALIZED SCORING] Running controller for challenge: {challenge}"
+        )
+        bt.logging.info(
+            f"[CENTRALIZED SCORING] Going to score {len(new_commits)} commits for challenge: {challenge}"
+        )
+        # This challenge controll will run with new inputs and reference commit input
+        # Reference commits are collected from yesterday, so if same docker_hub_id commited same day, they can share comparison_logs field, and of course, scoring_logs field
+        # If same docker_hub_id commited different day, the later one expected to be ignored anyway
+        controller = self.active_challenges[challenge]["controller"](
+            challenge_name=challenge,
+            miner_commits=new_commits,
+            reference_comparison_commits=unique_commits_cached_data,
+            challenge_info=self.active_challenges[challenge],
+            seed_inputs=seed_inputs,
+        )
+        # Run challenge controller, the controller update commit 's scoring logs and reference comparison logs directly
+        controller.start_challenge()
 
-    def _fetch_miner_submit(self):
+        # 4. Do comparison for new commits with each other, we only compare with reference commits
+        self._compare_miner_commits(
+            challenge=challenge,
+            revealed_commits_list=new_commits,
+            compare_with_each_other=False,
+        )
+
+    def _compare_miner_commits(
+        self,
+        challenge: str,
+        revealed_commits_list: list[MinerChallengeCommit],
+        compare_with_each_other: bool = False,
+    ):
         """
-        Fetch miner submit from each validator on storage and aggregate them
+        Compare miner commits for similarity checking.
+
+        Args:
+            challenge (str): Challenge name
+            revealed_commits_list (list[MinerChallengeCommit]): Commits to compare
+            compare_with_each_other (bool): If True, compares all commits with each other
+                                          If False, only compares with reference commits
+
+        Note: Used in two contexts:
+        1. During regular scoring: compare_with_each_other=False to compare with reference commits
+        2. At scoring hour: compare_with_each_other=True to compare all commits submitted within the same day with each other
         """
-        active_validator_uids = []
+        if not revealed_commits_list:
+            bt.logging.info(
+                f"[CENTRALIZED SCORING] No commits for challenge: {challenge}, skipping"
+            )
+            return
 
-        for uid, stake in enumerate(self.metagraph.S):
-            if stake > constants.MIN_VALIDATOR_STAKE:
-                active_validator_uids.append(uid)
+        bt.logging.info(
+            f"[CENTRALIZED SCORING] Running comparer for challenge: {challenge}"
+        )
+        comparer = self.active_challenges[challenge]["comparer"](
+            challenge_name=challenge,
+            challenge_info=self.active_challenges[challenge],
+            miner_commits=revealed_commits_list,
+            compare_with_each_other=compare_with_each_other,
+        )
+        # Run comparison, the comparer update commit 's penalty and comparison logs directly
+        comparer.start_comparison()
 
-        # Fetch all miner submit from each validator
-        validator_miner_submit = {} # key: validator_uid, value: dict of miner_submit for that validator, map from miner_ss58_address to miner_submit of each challenge
-        for validator_uid in active_validator_uids:
+        bt.logging.success(
+            f"[CENTRALIZED SCORING] Comparison for challenge: {challenge} has been completed"
+        )
+
+    def run(self):
+        bt.logging.info("Starting RewardApp loop.")
+        # Try set weights after initial sync
+        try:
+            bt.logging.info("Initializing weights")
+            self.set_weights()
+        except Exception:
+            bt.logging.error(f"Initial set weights error: {traceback.format_exc()}")
+
+        while True:
+            start_epoch = time.time()
+
             try:
-                endpoint = constants.STORAGE_URL + "/fetch-miner-submit"
+                self.forward()
+            except Exception as e:
+                bt.logging.error(f"Forward error: {traceback.format_exc()}")
+
+            end_epoch = time.time()
+            elapsed = end_epoch - start_epoch
+            time_to_sleep = max(0, self.config.reward_app.epoch_length - elapsed)
+            bt.logging.info(f"Epoch finished. Sleeping for {time_to_sleep} seconds.")
+
+
+            try:
+                self.set_weights()
+            except Exception:
+                bt.logging.error(f"Set weights error: {traceback.format_exc()}")
+
+            try:
+                self.resync_metagraph()
+            except Exception:
+                bt.logging.error(f"Resync metagraph error: {traceback.format_exc()}")
+
+            except KeyboardInterrupt:
+                bt.logging.success("Keyboard interrupt detected. Exiting validator.")
+                exit()
+
+    # MARK: Commit Management
+    def _update_validators_miner_commits(self):
+        """
+        Fetch all miner commits for challenges from all valid validators in the subnet.
+
+        Process:
+        1. Filter valid validators based on minimum stake requirement
+        2. For each valid validator:
+           - Fetch their latest miner commits from storage
+           - Validate and process commits for active miners
+           - Store in self.validators_miner_commits with validator (uid, hotkey) as key
+        """
+        # Get list of valid validators based on stake
+        valid_validators = []
+        for validator_uid, validator_ss58_address in enumerate(self.metagraph.hotkeys):
+            stake = self.metagraph.S[validator_uid]
+            if stake >= constants.MIN_VALIDATOR_STAKE:
+                valid_validators.append((validator_uid, validator_ss58_address))
+
+        bt.logging.info(
+            f"[CENTRALIZED COMMIT UPDATES] Found {len(valid_validators)} valid validators"
+        )
+
+        # Initialize/clear validators_miner_commits for this round
+        self.validators_miner_commits = {}
+
+        for validator_uid, validator_hotkey in valid_validators:
+            # Skip if request fails
+            try:
+                endpoint = f"{constants.STORAGE_URL}/fetch-latest-miner-commits"
                 data = {
                     "validator_uid": validator_uid,
-                    "validator_ss58_address": self.metagraph.hotkeys[validator_uid],
-                    "challenge_names": list(self.active_challenges.keys())
+                    "validator_hotkey": validator_hotkey,
+                    "challenge_names": list(self.active_challenges.keys()),
                 }
-                response = requests.post(endpoint, json=data)
+                response = requests.post(
+                    endpoint, headers=self.validator_request_header_fn(data), json=data
+                )
+                response.raise_for_status()
+                # Only continue if response is successful
+                data = response.json()
+                this_validator_miner_commits: dict[
+                    tuple[int, str], dict[str, MinerChallengeCommit]
+                ] = {}
+                # Process miner submissions for this validator
+                for miner_hotkey, miner_commits in data["miner_commits"].items():
+                    if miner_hotkey not in self.metagraph.hotkeys:
+                        # Skip if miner hotkey is not in metagraph
+                        continue
 
-                if response.status_code == 200:
-                    data = response.json()
+                    for challenge_name, miner_commit in miner_commits.items():
+                        miner_commit = MinerChallengeCommit.model_validate(miner_commit)
 
-                    for miner_ss58_address, challenges in data["miner_submit"].items():
-                        if miner_ss58_address in self.metagraph.hotkeys:
-                            miner_uid = self.metagraph.hotkeys.index(miner_ss58_address)
-                        else:
-                            # Skip if miner ss58_address no longer in metagraph
-                            continue
-                        for challenge_name, commit_data in challenges.items():
-                            validator_miner_submit.setdefault(validator_uid, {}).setdefault(miner_ss58_address, {})[challenge_name] = {
-                                "commit_timestamp": commit_data["commit_timestamp"],
-                                "encrypted_commit": commit_data["encrypted_commit"],
-                                "key": commit_data["key"],
-                                "commit": commit_data["commit"],
-                                "uid": miner_uid,
-                                "ss58_address": miner_ss58_address
-                            }
+                        this_validator_miner_commits.setdefault(
+                            (miner_commit.miner_uid, miner_commit.miner_hotkey), {}
+                        )[miner_commit.challenge_name] = miner_commit
 
-                    print(f"[SUCCESS] Fetched miner submit data from storage for validator {validator_uid}.")
+                self.validators_miner_commits[(validator_uid, validator_hotkey)] = (
+                    this_validator_miner_commits
+                )
+                bt.logging.success(
+                    f"[CENTRALIZED COMMIT UPDATES] Fetched miner commits data from validator {validator_uid}, hotkey: {validator_hotkey}"
+                )
+
+            except Exception:
+                bt.logging.warning(
+                    f"[CENTRALIZED COMMIT UPDATES] Failed to fetch data for validator {validator_uid}, hotkey: {validator_hotkey}: {traceback.format_exc()}"
+                )
+                continue
+
+        bt.logging.success(
+            f"[CENTRALIZED COMMIT UPDATES] Updated validators_miner_submit with data from {len(self.validators_miner_commits)} validators"
+        )
+
+    def _update_miner_commits(self):
+        """
+        Aggregate miner commits from all validators into a single state.
+
+        Process:
+        1. Create new aggregated state from all validator commits
+        2. For each miner/challenge:
+           - Keep latest commit based on timestamp
+           - For same encrypted_commit:
+             * Use older commit timestamp
+             * Preserve key and commit information
+           - For different encrypted_commit:
+             * Keep newer one based on timestamp
+        3. Merge scoring data from existing state for unchanged commits
+        4. Update self.miner_commits_cache for quick lookups
+        """
+        # Create new miner commits dict for aggregation
+        new_miner_commits: dict[tuple[int, str], dict[str, MinerChallengeCommit]] = {}
+
+        # Aggregate commits from all validators
+        for _, miner_commits_from_validator in self.validators_miner_commits.items():
+            for (
+                miner_uid,
+                miner_hotkey,
+            ), miner_commits_in_challenges in miner_commits_from_validator.items():
+                if miner_hotkey not in self.metagraph.hotkeys:
+                    # Skip if miner hotkey is not in metagraph
+                    continue
+
+                miner_key = (miner_uid, miner_hotkey)
+
+                # Initialize if first time seeing this miner
+                if miner_key not in new_miner_commits:
+                    new_miner_commits[miner_key] = miner_commits_in_challenges
                 else:
-                    print(f"[ERROR] Failed to fetch miner submit data: {response.status_code} - {response.text}")
-            except Exception as e:
-                print(f"[ERROR] Error fetching miner submit data from storage for validator {validator_uid}: {e}")
-
-        # Aggregate miner submits from validators
-        miner_submit = {}
-        for validator_uid, this_validator_miner_submit in validator_miner_submit.items():
-            for miner_ss58_address, miner_submissions in this_validator_miner_submit.items():
-                if miner_ss58_address not in miner_submit:
-                    # Update if first met this miner_ss58_address
-                    miner_submit[miner_ss58_address] = this_validator_miner_submit[miner_ss58_address]
-                else:
-                    for challenge_name, commit_data in miner_submissions.items():
-                        if challenge_name not in miner_submit[miner_ss58_address]:
-                            miner_submit[miner_ss58_address][challenge_name] = commit_data
+                    # Update miner commits
+                    for (
+                        challenge_name,
+                        miner_commit,
+                    ) in miner_commits_in_challenges.items():
+                        if challenge_name not in new_miner_commits[miner_key]:
+                            new_miner_commits[miner_key][challenge_name] = miner_commit
                         else:
-                            current_miner_submit_for_challenge = miner_submit[miner_ss58_address][challenge_name]
-                            if commit_data["encrypted_commit"] == current_miner_submit_for_challenge["encrypted_commit"]:
+                            current_miner_commit = new_miner_commits[miner_key][
+                                challenge_name
+                            ]
+                            if (
+                                miner_commit.encrypted_commit
+                                == current_miner_commit.encrypted_commit
+                            ):
                                 # If encrypted commit is the same, we update to older commit timestamp and add unknown commit and key field if possible
-                                miner_submit[miner_ss58_address][challenge_name]["commit_timestamp"] = commit_data["commit_timestamp"]
-                                if not current_miner_submit_for_challenge["key"]:
-                                    miner_submit[miner_ss58_address][challenge_name]["key"] = commit_data["key"]
-                                if not current_miner_submit_for_challenge["commit"]:
-                                    miner_submit[miner_ss58_address][challenge_name]["commit"] = commit_data["commit"]
+                                if (
+                                    miner_commit.commit_timestamp
+                                    and current_miner_commit.commit_timestamp
+                                    and miner_commit.commit_timestamp
+                                    < current_miner_commit.commit_timestamp
+                                ):
+                                    # Update to older commit timestamp
+                                    current_miner_commit.commit_timestamp = (
+                                        miner_commit.commit_timestamp
+                                    )
+                                if not current_miner_commit.key:
+                                    # Add unknown key if possible
+                                    current_miner_commit.key = miner_commit.key
+                                if not current_miner_commit.commit:
+                                    # Add unknown commit if possible
+                                    current_miner_commit.commit = miner_commit.commit
                             else:
                                 # If encrypted commit is different, we compare commit timestamp
-                                if commit_data["commit_timestamp"] > current_miner_submit_for_challenge["commit_timestamp"]:
+                                if (
+                                    miner_commit.commit_timestamp
+                                    and current_miner_commit.commit_timestamp
+                                    and miner_commit.commit_timestamp
+                                    > current_miner_commit.commit_timestamp
+                                ):
                                     # If newer commit timestamp, update to the latest commit
-                                    miner_submit[miner_ss58_address][challenge_name] = commit_data
+                                    current_miner_commit.commit_timestamp = (
+                                        miner_commit.commit_timestamp
+                                    )
                                 else:
                                     # If older commit timestamp, skip
                                     continue
 
-        self.miner_submit = miner_submit
+        # Merge scoring data from existing state
+        for miner_key, existing_challenges in self.miner_commits.items():
+            if miner_key not in new_miner_commits:
+                continue
 
-    def fetch_submission_scoring_logs(self, challenge_names: list):
-        endpoint = constants.STORAGE_URL + "/fetch-centralized-score"
+            for challenge_name, existing_commit in existing_challenges.items():
+                if challenge_name not in new_miner_commits[miner_key]:
+                    continue
 
-        submission_scoring_logs = {}
-        try:
-            response = requests.post(endpoint, json={"challenge_names": challenge_names})
-            if response.status_code == 200:
-                print("[SUCCESS] Submission scoring logs successfully fetched from storage.")
-                logs = response.json()["data"]
-                for log in logs:
-                    challenge_name = log["challenge_name"]
-                    docker_hub_id = log["docker_hub_id"]
-                    if challenge_name not in submission_scoring_logs:
-                        submission_scoring_logs[challenge_name] = {}
-                    for l in log["logs"]:
-                        l = self._normalize_log(l)
-                    submission_scoring_logs[challenge_name][docker_hub_id] = log["logs"]
+                new_commit = new_miner_commits[miner_key][challenge_name]
+                # If same encrypted commit, preserve stateful fields
+                if existing_commit.encrypted_commit == new_commit.encrypted_commit:
+                    new_commit.scoring_logs = existing_commit.scoring_logs
+                    new_commit.comparison_logs = existing_commit.comparison_logs
+                    new_commit.score = existing_commit.score
+                    new_commit.penalty = existing_commit.penalty
+                    new_commit.accepted = existing_commit.accepted
+        self.miner_commits = new_miner_commits
 
-            else:
-                print(f"[ERROR] Failed to fetch submission scoring logs from storage: {response.status_code} - {response.text}")
-            return submission_scoring_logs
-        except Exception as e:
-            print(f"[ERROR] Error fetching submission scoring logs from storage: {e}")
-            return {}
-
-    def save_submission_scoring_logs(self, new_submission_scoring_logs: dict):
-        endpoint = constants.STORAGE_URL + "/upload-centralized-score"
-        try:
-            # If all submission scoring logs are empty, return False
-            if all(not value for value in new_submission_scoring_logs.values()):
-                return False
-            # if all(not value for value in self.submission_scoring_logs.values()):
-            #     return False
-            # # If submission scoring logs are not updated, return False
-            # if self.previous_submission_scoring_logs == self.submission_scoring_logs:
-            #     return False
-            response = requests.post(endpoint, json={"data": new_submission_scoring_logs})
-            if response.status_code == 200:
-                print("[SUCCESS] Submission scoring logs successfully saved to storage.")
-                self.previous_submission_scoring_logs = copy.deepcopy(self.submission_scoring_logs)
-            else:
-                print(f"[ERROR] Failed to save submission scoring logs to storage: {response.status_code} - {response.text}")
-        except Exception as e:
-            print(f"[ERROR] Error saving submission scoring logs to storage: {e}")
-        return True
-
-    def get_scoring_logs(self, challenge_name: str):
-        challenge_logs = self.submission_scoring_logs.get(challenge_name, {})
-
-        # Filter logs to only include docker IDs that are in mapping_docker_id_miner_id
-        filtered_logs = {
-            docker_id: logs
-            for docker_id, logs in challenge_logs.items()
-            if docker_id in self.mapping_docker_id_miner_id
+        # Sort by UID to make sure all next operations are order consistent
+        self.miner_commits = {
+            (uid, ss58_address): commits
+            for (uid, ss58_address), commits in sorted(
+                self.miner_commits.items(), key=lambda item: item[0]
+            )
         }
+
+        # Update miner commits cache
+        self.miner_commits_cache = {
+            f"{commit.challenge_name}---{commit.encrypted_commit}": commit
+            for _, commits in self.miner_commits.items()
+            for commit in commits.values()
+        }
+
+    # MARK: Storage
+    def _store_centralized_scoring(self, challenge_name: str = None):
+        """
+        Store scoring results to centralized storage.
+
+        Args:
+            challenge_name (str, optional): Specific challenge to store.
+                                          If None, stores all challenges.
+
+        Stores:
+            - Challenge name
+            - Docker hub ID
+            - Scoring logs
+            - Comparison logs
+        """
+        challenge_names = (
+            [challenge_name] if challenge_name else list(self.scoring_results.keys())
+        )
+        endpoint = f"{constants.STORAGE_URL}/upload-centralized-score"
+
+        all_scoring_results = []
+
+        for challenge_name in challenge_names:
+            scoring_results_to_send: list[dict] = []
+            # Send batch of maximum 5 results at a time to avoid huge payload
+            for docker_hub_id, result in self.scoring_results.get(challenge_name, {}).items():
+                scoring_result = {
+                    "challenge_name": challenge_name,
+                    "docker_hub_id": docker_hub_id,
+                    "scoring_logs": [
+                        scoring_log.model_dump()
+                        for scoring_log in result.get("scoring_logs", [])
+                    ],
+                    "comparison_logs": {
+                        docker_hub_id: [
+                            comparison_log.model_dump()
+                            for comparison_log in _comparison_logs
+                        ]
+                        for docker_hub_id, _comparison_logs in result.get(
+                            "comparison_logs", {}
+                        ).items()
+                    }
+                }
+                scoring_results_to_send.append(scoring_result)
+                all_scoring_results.append(scoring_result)
+
+                if len(scoring_results_to_send) >= 5:
+                    try:
+                        data = {
+                            "scoring_results": scoring_results_to_send
+                        }
+                        response = requests.post(
+                            endpoint, headers=self.validator_request_header_fn(data), json=data
+                        )
+                        response.raise_for_status()
+                        scoring_results_to_send = []
+                    except Exception:
+                        bt.logging.error(f"Failed to send scoring results to storage: {traceback.format_exc()}")
+                        scoring_results_to_send = []
+
+            if scoring_results_to_send:
+                try:
+                    data = {
+                        "scoring_results": scoring_results_to_send
+                    }
+                    response = requests.post(
+                        endpoint, headers=self.validator_request_header_fn(data), json=data
+                    )
+                    response.raise_for_status()
+                except Exception:
+                    bt.logging.error(f"Failed to send scoring results to storage: {traceback.format_exc()}")
+
+        if all_scoring_results:
+            try:
+                with open("scoring_results.json", "w") as f:
+                    json.dump(all_scoring_results, f)
+            except Exception:
+                bt.logging.error(f"Failed to save scoring results to file: {traceback.format_exc()}")
+
+    def _fetch_centralized_scoring(
+        self, challenge_names: list[str] = []
+    ) -> dict[str, dict[str, Union[list[ScoringLog], dict[str, ComparisonLog]]]]:
+        """
+        Fetch scoring results from centralized storage.
+
+        Args:
+            challenge_names (list[str]): List of challenges to fetch.
+                                       If empty, fetches all active challenges.
+
+        Returns:
+            dict: Mapping of {challenge_name: {docker_hub_id: {scoring_logs, comparison_logs}}}
+        """
+        if not challenge_names:
+            # Fetch all challenges
+            challenge_names = list(self.active_challenges.keys())
+
+        endpoint = f"{constants.STORAGE_URL}/fetch-centralized-score"
+        data = {
+            "challenge_names": challenge_names,
+        }
+        response = requests.post(
+            endpoint, headers=self.validator_request_header_fn(data), json=data
+        )
+        response.raise_for_status()
+        data = response.json()["data"]
+
+        scoring_results = {}
+        for results in data:
+            scoring_results.setdefault(results["challenge_name"], {})[
+                results["docker_hub_id"]
+            ] = {
+                "scoring_logs": [
+                    ScoringLog.model_validate(scoring_log)
+                    for scoring_log in results["scoring_logs"]
+                ],
+                "comparison_logs": {
+                    docker_hub_id: [
+                        ComparisonLog.model_validate(comparison_log)
+                        for comparison_log in _comparison_logs
+                    ]
+                    for docker_hub_id, _comparison_logs in results[
+                        "comparison_logs"
+                    ].items()
+                },
+            }
+        return scoring_results
+
+    def _sync_scoring_results_from_storage_to_cache(self):
+        """
+        Sync scoring results (self.scoring_results) from storage to cache.
+        This method will update the cache with scoring results from storage.
+        It will also delete cache entries that are not in self.scoring_results.
+        """
+        # Iter all the keys in all cache corespond to active challenges
+        for challenge_name in self.active_challenges.keys():
+            cache = self.storage_manager._get_cache(challenge_name)
+            cache_keys_to_delete = []
+
+            for hashed_cache_key in cache.iterkeys():
+                commit = cache.get(hashed_cache_key)
+                try:
+                    commit = MinerChallengeCommit.model_validate(
+                        commit
+                    )  # Model validate the commit
+                except Exception:
+                    # Skip if commit is not valid
+                    # Do this if we want to clean up invalid commits
+                    # cache_keys_to_delete.append(hashed_cache_key)
+                    continue
+
+                # Check if docker_hub_id is in self.scoring_results
+                if commit.docker_hub_id not in self.scoring_results[challenge_name]:
+                    # Do this if we want to clean up commits with no scoring results
+                    # cache_keys_to_delete.append(hashed_cache_key)
+                    continue
+
+                # Found the commit in self.scoring_results, now we make sure cache have correct scoring_logs
+                if not commit.scoring_logs:
+                    # If not scoring_logs, we add the scoring_logs from self.scoring_results
+                    commit.scoring_logs = self.scoring_results[challenge_name][
+                        commit.docker_hub_id
+                    ]["scoring_logs"]
+                    cache[hashed_cache_key] = commit.model_dump()
+                else:
+                    # Check for each entries in scoring_logs for miner_input and miner_output, they should not be None
+                    scoring_logs_with_none = []
+                    for scoring_log in commit.scoring_logs:
+                        if (
+                            scoring_log.miner_input is None
+                            or scoring_log.miner_output is None
+                        ):
+                            scoring_logs_with_none.append(scoring_log)
+
+                    # If there are any scoring logs with None, we use the scoring_logs from self.scoring_results and update the cache
+                    if any(scoring_logs_with_none):
+                        commit.scoring_logs = self.scoring_results[challenge_name][
+                            commit.docker_hub_id
+                        ]["scoring_logs"]
+                        cache[hashed_cache_key] = commit.model_dump()
+
+            # Clean up cache entries that we want to delete
+            for hashed_cache_key in cache_keys_to_delete:
+                cache.delete(hashed_cache_key)
+
+    # MARK: Endpoints
+    async def get_scoring_result(
+        self,
+        challenge_name: Annotated[str, Body(..., embed=True)],
+        encrypted_commits: Annotated[list[str], Body(..., embed=True)],
+    ):
+        """
+        API endpoint to get scoring logs for specific docker hub IDs.
+        This method check for encrypted_commit in miner_commits and return the cached commit result.
+
+        Args:
+            docker_hub_ids: List of docker hub IDs to look up
+
+        Returns:
+            dict: {
+                docker_hub_id: scoring_log or None if not found
+            }
+        """
+        assert challenge_name in self.active_challenges, (
+            f"Challenge {challenge_name} is not active"
+        )
+
+        results: dict[str, MinerChallengeCommit] = {}
+
+        for encrypted_commit in encrypted_commits:
+            # Try in-memory cache first
+            cache_key = f"{challenge_name}---{encrypted_commit}"
+            commit_result = self.miner_commits_cache.get(cache_key, None)
+            if commit_result:
+                results[encrypted_commit] = commit_result.public_view()
+                continue
+
+            # Fallback to disk cache
+            try:
+                hashed_cache_key = self.storage_manager.hash_cache_key(encrypted_commit)
+                challenge_cache = self.storage_manager._get_cache(challenge_name)
+                cached_data = challenge_cache.get(hashed_cache_key)
+
+                if cached_data:
+                    commit_result = MinerChallengeCommit.model_validate(cached_data)
+                    results[encrypted_commit] = commit_result.public_view()
+                else:
+                    results[encrypted_commit] = None
+            except Exception as e:
+                bt.logging.error(f"Error retrieving from disk cache: {e}")
+                results[encrypted_commit] = None
 
         return {
-            "submission_scoring_logs": filtered_logs,
-            "is_scoring_done": self.is_scoring_done.get(challenge_name, False)
+            "status": "success",
+            "message": "Scoring results retrieved successfully",
+            "data": {
+                "commits": results,
+                "is_done": self.is_scoring_done.get(challenge_name),
+            },
         }
 
-    def _sync_metagraph(self, sync_interval= 60 * 10):
-        """Background thread to sync metagraph."""
-        while True:
-            time.sleep(sync_interval)
-            try:
-                self.metagraph.sync(lite=True)
-                bt.logging.success("Metagraph synced successfully.")
-            except Exception as e:
-                bt.logging.error(f"Error syncing metagraph: {e}")
-
-    def _init_challenge_records_from_subnet(self, is_today_scored: bool = False):
-        try:
-            endpoint = constants.STORAGE_URL + "/fetch-challenge-records"
-            data = {
-                "validator_ss58_address": self.REWARD_APP_SERVER_SS58_ADDRESS,
-                "is_today_scored": is_today_scored,
-                "challenge_names": list(self.active_challenges.keys())
-            }
-            response = requests.post(endpoint, json=data)
-
-            if response.status_code == 200:
-                data = response.json()
-
-                for challenge_name, challenge_record in data.items():
-                    if challenge_name in self.miner_managers:
-                        self.miner_managers[challenge_name].challenge_records = {date: ChallengeRecord(**record) for date, record in challenge_record.items()}
-                bt.logging.success("[INIT] Challenge records data successfully initialized from storage.")
-            else:
-                bt.logging.error(f"[INIT] Failed to fetch challenge records data: {response.status_code} - {response.text}")
-        except Exception as e:
-            bt.logging.error(f"[INIT] Error initializing challenge records data from storage: {e}")
-            raise  # Re-raise to handle initialization failure
-
-    def _store_challenge_records(self):
-        challenge_records = {}
-        for challenge_name, miner_manager in self.miner_managers.items():
-            challenge_records[challenge_name] =  {
-                date: record.__dict__ for date, record in miner_manager.challenge_records.items()
-            }
-        data = {
-            "validator_ss58_address": self.REWARD_APP_SERVER_SS58_ADDRESS,
-            "validator_uid": self.REWARD_APP_SERVER_UID,
-            "challenge_records": challenge_records,
-            "signature": self.REWARD_APP_KEY,
-            "nonce": str(time.time_ns())
-        }
-        self.storage_manager.update_challenge_records(data)
-        print("[SUCCESS] Challenge records successfully stored to storage.")
-
-    def _normalize_log(self, log: dict):
-        if type(log.get("score")) == int:
-            log["score"] = float(log["score"])
-        elif not type(log.get("score")) == float:
-            log["score"] = 0
-        return log
 
 if __name__ == "__main__":
-    bt.logging.enable_info()
-    args = get_args()
-    app = RewardApp(args)
-
-    uvicorn.run(
-        app.app,
-        host="0.0.0.0",
-        port=args.port,
-    )
+    # Initialize and run app
+    with RewardApp(get_reward_app_config()) as app:
+        while True:
+            bt.logging.info("RewardApp is running...")
+            time.sleep(constants.EPOCH_LENGTH // 4)
